* WIP beginning to decompose suite creation
