+++ .mypy.ini
@@ -1,0 +1,12 @@
+# Global options:
+
+[mypy]
+python_version = 2.7
+warn_return_any = True
+warn_unused_configs = True
+ignore_missing_imports = True
+
+# Per-module options:
+
+[mypy-great_expectations.render.renderer.suite_scaffold_notebook_renderer]
+disallow_untyped_defs = True
